当たり判定が機能するように行った変更、それに伴ったスクロール方法の変更

+++ src/main/java/jp/u_donbei/udon_pasuta/object/Movable.java
@@ -8,10 +8,10 @@ public interface Movable {
 	 * X方向に動く。
 	 * @param x X方向の差分
 	 */
-	void moveX(int x);
+	void moveX(double x);
 	/**
 	 * Y方向に動く。
 	 * @param y Y方向の差分
 	 */
-	void moveY(int y);
+	void moveY(double y);
 }
